initial 1

+++ pages/api/post/[id]/index.js
@@ -1,19 +1,28 @@
-import db from "../../../../utils/db";
-import postModel from "../../../../models/post";
+import db from '../../../../utils/db';
+import postModel from '../../../../models/post';
 
-import { getSession } from "next-auth/react";
+import { getSession } from 'next-auth/react';
 
 export default async (req, res) => {
   switch (req.method) {
-    case "GET":
+    case 'GET':
       await getPost(req, res);
       break;
-    case "PUT":
+
+    case 'PUT':
       await editPost(req, res);
       break;
 
-    case "DELETE":
+    case 'DELETE':
       await deletePost(req, res);
+      break;
+
+    case 'PUT':
+      await likePost(req, res);
+      break;
+
+    case 'PUT':
+      await dislikePost(req, res);
       break;
   }
 };
@@ -35,12 +44,8 @@ export const editPost = async (req, res) => {
   const session = await getSession({ req });
 
   if (!session) {
-    return res.status(401).send("you are not authenticated");
+    return res.status(401).send('you are not authenticated');
   }
-  //   const userId = session.user._id;
-
-  //   if (condition) {
-  //   }
 
   try {
     await db.connect();
@@ -50,38 +55,71 @@ export const editPost = async (req, res) => {
     const data = await postModel.findByIdAndUpdate(id, updatedData, options);
 
     await db.disconnect();
-    res.status(200).json({ message: "updated successfully", data });
+    res.status(200).json({ message: 'updated successfully', data });
   } catch (error) {
     await db.disconnect();
-    res.status(500).json({ message: "something went wrong" });
+    res.status(500).json({ message: 'something went wrong' });
   }
-  // const userId = session.user._id;
-
-  // const post = req.body;
-
-  // await db.connect();
-
-  // const newPost = new postModel({
-  // 	...post,
-  // 	user: userId
-  // });
-
-  // try {
-  // 	await newPost.save();
-
-  // 	await db.disconnect();
-  // 	res.status(201).json(JSON.stringify(newPost));
-  // } catch (error) {
-  // 	res.status(409).json({ message: error });
-  // }
 };
 
 export const deletePost = async (req, res) => {
   const id = req.query.id;
 
+  const session = await getSession({ req });
+
+  if (!session) {
+    return res.status(401).send('you are not authenticated');
+  }
+
   try {
     await postModel.findByIdAndRemove(id);
-    res.json({ message: "Post deleted successfully." });
+    res.json({ message: 'Post deleted successfully.' });
+  } catch (error) {
+    res.status(409).json({ message: error });
+  }
+};
+
+export const likePost = async (req, res) => {
+  const session = await getSession({ req });
+
+  if (!session) {
+    return res.status(401).send('you are not authenticated');
+  }
+
+  const userId = session.user._id;
+
+  const id = req.query.id;
+
+  try {
+    await postModel.findByIdAndUpdate(id, {
+      $addToSet: { likes: userId },
+      $pull: { dislikes: userId },
+    });
+
+    res.json({ message: 'the vido has been liked' });
+  } catch (error) {
+    res.status(409).json({ message: error });
+  }
+};
+
+export const dislikePost = async (req, res) => {
+  const session = await getSession({ req });
+
+  if (!session) {
+    return res.status(401).send('you are not authenticated');
+  }
+
+  const userId = session.user._id;
+
+  const id = req.query.id;
+
+  try {
+    await postModel.findByIdAndUpdate(id, {
+      $addToSet: { dislikes: userId },
+      $pull: { likes: userId },
+    });
+
+    res.json({ message: 'the vido has been disliked' });
   } catch (error) {
     res.status(409).json({ message: error });
   }
